refactor(settings): group badge components into a folder
Move the four badge settings components (Badges, BadgeForm, BadgeAssignments,
BadgeAssignmentForm) into Settings/Badges/, matching the per-feature subfolder
convention used by Coupons/ and Transactions/ here and by CRM/Helpdesk. Pure
move plus the import-path updates; no behaviour change.
