+++ frontend/src/components/Settings/Badges/BadgeAssignmentForm.vue
