+++ frontend/src/components/Settings/Badges/BadgeForm.vue
